requirements.txt
pandas
sqlalchemy
pymysql
azure-storage-blob
python-dotenv
streamlit
pyodbc